refactor thread_worker stuff
